+++ drivers/net/wireless/ath/ath9k/ath9k.h
@@ -1013,7 +1013,7 @@ struct ath_softc {
 	ktime_t last_event_time;
 #endif
 
-	unsigned long driver_data;
+	uintptr_t driver_data;
 
 	u8 gtt_cnt;
 	u32 intrstatus;
